+++ packages/theme/i18n/es.js
@@ -1,7 +1,7 @@
 
 i18n.map('es', {
 	website_name: "KeplerJs",
-	website_slogan: "The open source full-stack geosocial network framework",
+	website_slogan: "The open source full-stack geosocial network platform",
 
 	map_zoomout: " ",
 	map_zoomin: " ",